refactor(cli): extract sandbox create stream from onboard.js
