+++ bin/lib/onboard.js
@@ -65,6 +65,7 @@ const dashboard = require("../../dist/lib/dashboard");
 const httpProbe = require("../../dist/lib/http-probe");
 const modelPrompts = require("../../dist/lib/model-prompts");
 const providerModels = require("../../dist/lib/provider-models");
+const sandboxCreateStream = require("../../dist/lib/sandbox-create-stream");
 const validationRecovery = require("../../dist/lib/validation-recovery");
 const webSearch = require("../../dist/lib/web-search");
 
@@ -223,211 +224,7 @@ function repairRecordedSandbox(sandboxName) {
   registry.removeSandbox(sandboxName);
 }
 
-function streamSandboxCreate(command, env = process.env, options = {}) {
-  const child = spawn("bash", ["-lc", command], {
-    cwd: ROOT,
-    env,
-    stdio: ["ignore", "pipe", "pipe"],
-  });
-
-  const lines = [];
-  let pending = "";
-  let lastPrintedLine = "";
-  let sawProgress = false;
-  let settled = false;
-  let polling = false;
-  const pollIntervalMs = options.pollIntervalMs || 2000;
-  const heartbeatIntervalMs = options.heartbeatIntervalMs || 5000;
-  const silentPhaseMs = options.silentPhaseMs || 15000;
-  const startedAt = Date.now();
-  let lastOutputAt = startedAt;
-  let currentPhase = "build";
-  let lastHeartbeatPhase = null;
-  let lastHeartbeatBucket = -1;
-
-  function getDisplayWidth() {
-    return Math.max(60, Number(process.stdout.columns || 100));
-  }
-
-  function trimDisplayLine(line) {
-    const width = getDisplayWidth();
-    const maxLen = Math.max(40, width - 4);
-    if (line.length <= maxLen) return line;
-    return `${line.slice(0, Math.max(0, maxLen - 3))}...`;
-  }
-
-  function printProgressLine(line) {
-    const display = trimDisplayLine(line);
-    if (display !== lastPrintedLine) {
-      console.log(display);
-      lastPrintedLine = display;
-    }
-  }
-
-  function elapsedSeconds() {
-    return Math.max(0, Math.floor((Date.now() - startedAt) / 1000));
-  }
-
-  function setPhase(nextPhase) {
-    if (!nextPhase || nextPhase === currentPhase) return;
-    currentPhase = nextPhase;
-    lastHeartbeatPhase = null;
-    lastHeartbeatBucket = -1;
-    const phaseLine =
-      nextPhase === "build"
-        ? "  Building sandbox image..."
-        : nextPhase === "upload"
-          ? "  Uploading image into OpenShell gateway..."
-          : nextPhase === "create"
-            ? "  Creating sandbox in gateway..."
-            : nextPhase === "ready"
-              ? "  Waiting for sandbox to become ready..."
-              : null;
-    if (phaseLine) printProgressLine(phaseLine);
-  }
-
-  function finish(result) {
-    if (settled) return;
-    settled = true;
-    if (pending) flushLine(pending);
-    if (readyTimer) clearInterval(readyTimer);
-    if (heartbeatTimer) clearInterval(heartbeatTimer);
-    resolvePromise(result);
-  }
-
-  function detachChild() {
-    child.stdout?.removeAllListeners?.("data");
-    child.stderr?.removeAllListeners?.("data");
-    child.stdout?.destroy?.();
-    child.stderr?.destroy?.();
-    child.removeAllListeners?.("error");
-    child.removeAllListeners?.("close");
-    child.unref?.();
-  }
-
-  function shouldShowLine(line) {
-    return (
-      /^ {2}Building image /.test(line) ||
-      /^ {2}Step \d+\/\d+ : /.test(line) ||
-      /^ {2}Context: /.test(line) ||
-      /^ {2}Gateway: /.test(line) ||
-      /^Successfully built /.test(line) ||
-      /^Successfully tagged /.test(line) ||
-      /^ {2}Built image /.test(line) ||
-      /^ {2}Pushing image /.test(line) ||
-      /^\s*\[progress\]/.test(line) ||
-      /^ {2}Image .*available in the gateway/.test(line) ||
-      /^Created sandbox: /.test(line) ||
-      /^✓ /.test(line)
-    );
-  }
-
-  function flushLine(rawLine) {
-    const line = rawLine.replace(/\r/g, "").trimEnd();
-    if (!line) return;
-    lines.push(line);
-    lastOutputAt = Date.now();
-    if (/^ {2}Building image /.test(line) || /^ {2}Step \d+\/\d+ : /.test(line)) {
-      setPhase("build");
-    } else if (
-      /^ {2}Pushing image /.test(line) ||
-      /^\s*\[progress\]/.test(line) ||
-      /^ {2}Image .*available in the gateway/.test(line)
-    ) {
-      setPhase("upload");
-    } else if (/^Created sandbox: /.test(line)) {
-      setPhase("create");
-    }
-    if (shouldShowLine(line) && line !== lastPrintedLine) {
-      printProgressLine(line);
-      sawProgress = true;
-    }
-  }
-
-  function onChunk(chunk) {
-    pending += chunk.toString();
-    const parts = pending.split("\n");
-    pending = parts.pop();
-    parts.forEach(flushLine);
-  }
-
-  child.stdout.on("data", onChunk);
-  child.stderr.on("data", onChunk);
-
-  let resolvePromise;
-  const readyTimer = options.readyCheck
-    ? setInterval(() => {
-        if (settled || polling) return;
-        polling = true;
-        try {
-          let ready = false;
-          try {
-            ready = !!options.readyCheck();
-          } catch {
-            return;
-          }
-          if (!ready) return;
-          setPhase("ready");
-          const detail = "Sandbox reported Ready before create stream exited; continuing.";
-          lines.push(detail);
-          printProgressLine(`  ${detail}`);
-          try {
-            child.kill("SIGTERM");
-          } catch {
-            // Best effort only — the child may have already exited.
-          }
-          detachChild();
-          finish({ status: 0, output: lines.join("\n"), sawProgress: true, forcedReady: true });
-        } finally {
-          polling = false;
-        }
-      }, pollIntervalMs)
-    : null;
-  readyTimer?.unref?.();
-
-  setPhase("build");
-  const heartbeatTimer = setInterval(() => {
-    if (settled) return;
-    const silentForMs = Date.now() - lastOutputAt;
-    if (silentForMs < silentPhaseMs) return;
-    const elapsed = elapsedSeconds();
-    const bucket = Math.floor(elapsed / 15);
-    if (currentPhase === lastHeartbeatPhase && bucket === lastHeartbeatBucket) {
-      return;
-    }
-    const heartbeatLine =
-      currentPhase === "upload"
-        ? `  Still uploading image into OpenShell gateway... (${elapsed}s elapsed)`
-        : currentPhase === "create"
-          ? `  Still creating sandbox in gateway... (${elapsed}s elapsed)`
-          : currentPhase === "ready"
-            ? `  Still waiting for sandbox to become ready... (${elapsed}s elapsed)`
-            : `  Still building sandbox image... (${elapsed}s elapsed)`;
-    if (trimDisplayLine(heartbeatLine) !== lastPrintedLine) {
-      printProgressLine(heartbeatLine);
-      lastHeartbeatPhase = currentPhase;
-      lastHeartbeatBucket = bucket;
-    }
-  }, heartbeatIntervalMs);
-  heartbeatTimer.unref?.();
-
-  return new Promise((resolve) => {
-    resolvePromise = resolve;
-    child.on("error", (error) => {
-      // @ts-expect-error — Node ErrnoException has .code but TS types Error
-      const code = error && error.code;
-      const detail = code
-        ? `spawn failed: ${error.message} (${code})`
-        : `spawn failed: ${error.message}`;
-      lines.push(detail);
-      finish({ status: 1, output: lines.join("\n"), sawProgress: false });
-    });
-
-    child.on("close", (code) => {
-      finish({ status: code ?? 1, output: lines.join("\n"), sawProgress });
-    });
-  });
-}
+const { streamSandboxCreate } = sandboxCreateStream;
 
 function streamGatewayStart(command, env = process.env) {
   const child = spawn("bash", ["-lc", command], {
