+++ src/lib/sandbox-create-stream.test.ts
@@ -1,0 +1,97 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { EventEmitter } from "node:events";
+
+import { afterEach, describe, expect, it, vi } from "vitest";
+
+import { streamSandboxCreate } from "./sandbox-create-stream";
+
+class FakeReadable extends EventEmitter {
+  destroy() {}
+}
+
+class FakeChild extends EventEmitter {
+  stdout = new FakeReadable();
+  stderr = new FakeReadable();
+  kill = vi.fn();
+  unref = vi.fn();
+}
+
+describe("sandbox-create-stream", () => {
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it("streams visible progress lines and returns the collected output", async () => {
+    const child = new FakeChild();
+    const logLine = vi.fn();
+    const promise = streamSandboxCreate("echo create", process.env, {
+      logLine,
+      spawnImpl: () => child as never,
+      heartbeatIntervalMs: 1_000,
+      silentPhaseMs: 10_000,
+    });
+
+    child.stdout.emit(
+      "data",
+      Buffer.from("  Building image sandbox\n  Pushing image layers\nCreated sandbox: demo\n✓ Ready\n"),
+    );
+    child.emit("close", 0);
+
+    await expect(promise).resolves.toMatchObject({
+      status: 0,
+      sawProgress: true,
+      output: expect.stringContaining("Created sandbox: demo"),
+    });
+    expect(logLine).toHaveBeenCalledWith("  Building image sandbox");
+    expect(logLine).toHaveBeenCalledWith("  Pushing image layers");
+    expect(logLine).toHaveBeenCalledWith("Created sandbox: demo");
+  });
+
+  it("forces success when the sandbox becomes ready before the stream exits", async () => {
+    vi.useFakeTimers();
+
+    const child = new FakeChild();
+    let checks = 0;
+    const promise = streamSandboxCreate("echo create", process.env, {
+      spawnImpl: () => child as never,
+      readyCheck: () => {
+        checks += 1;
+        return checks >= 2;
+      },
+      pollIntervalMs: 5,
+      heartbeatIntervalMs: 1_000,
+      silentPhaseMs: 10_000,
+      logLine: vi.fn(),
+    });
+
+    child.stdout.emit("data", Buffer.from("  Building image sandbox\n"));
+    await vi.advanceTimersByTimeAsync(12);
+
+    await expect(promise).resolves.toMatchObject({
+      status: 0,
+      sawProgress: true,
+      forcedReady: true,
+      output: expect.stringContaining("Sandbox reported Ready before create stream exited"),
+    });
+    expect(child.kill).toHaveBeenCalledWith("SIGTERM");
+    expect(child.unref).toHaveBeenCalled();
+  });
+
+  it("reports spawn errors cleanly", async () => {
+    const child = new FakeChild();
+    const promise = streamSandboxCreate("echo create", process.env, {
+      spawnImpl: () => child as never,
+      logLine: vi.fn(),
+    });
+
+    child.emit("error", Object.assign(new Error("ENOENT"), { code: "ENOENT" }));
+
+    await expect(promise).resolves.toEqual({
+      status: 1,
+      output: "spawn failed: ENOENT (ENOENT)",
+      sawProgress: false,
+    });
+  });
+});
