+++ test/onboard.test.js
@@ -1299,18 +1299,23 @@ const { setupInference } = require(${onboardPath});
   });
 
   it("surfaces sandbox-create phases and silence heartbeats during long image operations", () => {
-    const source = fs.readFileSync(
+    const onboardSource = fs.readFileSync(
       path.join(import.meta.dirname, "..", "bin", "lib", "onboard.js"),
       "utf-8",
     );
+    const streamSource = fs.readFileSync(
+      path.join(import.meta.dirname, "..", "src", "lib", "sandbox-create-stream.ts"),
+      "utf-8",
+    );
 
-    assert.match(source, /function setPhase\(nextPhase\)/);
-    assert.match(source, /Building sandbox image\.\.\./);
-    assert.match(source, /Uploading image into OpenShell gateway\.\.\./);
-    assert.match(source, /Creating sandbox in gateway\.\.\./);
-    assert.match(source, /Still building sandbox image\.\.\. \(\$\{elapsed\}s elapsed\)/);
+    assert.match(onboardSource, /sandbox-create-stream/);
+    assert.match(streamSource, /function setPhase\(nextPhase/);
+    assert.match(streamSource, /Building sandbox image\.\.\./);
+    assert.match(streamSource, /Uploading image into OpenShell gateway\.\.\./);
+    assert.match(streamSource, /Creating sandbox in gateway\.\.\./);
+    assert.match(streamSource, /Still building sandbox image\.\.\. \(\$\{elapsed\}s elapsed\)/);
     assert.match(
-      source,
+      streamSource,
       /Still uploading image into OpenShell gateway\.\.\. \(\$\{elapsed\}s elapsed\)/,
     );
   });
